Split constants per device
