+++ switchbot/const/__init__.py
@@ -1,10 +1,11 @@
-"""Library to handle connection with Switchbot."""
+"""Switchbot Device Consts Library."""
 
 from __future__ import annotations
 
-from enum import Enum
+from ..enum import StrEnum
 
-from .enum import StrEnum
+# Preserve old LockStatus export for backwards compatibility
+from .lock import LockStatus as LockStatus
 
 DEFAULT_RETRY_COUNT = 3
 DEFAULT_RETRY_TIMEOUT = 1
@@ -61,13 +62,3 @@ class SwitchbotModel(StrEnum):
     RELAY_SWITCH_1PM = "Relay Switch 1PM"
     RELAY_SWITCH_1 = "Relay Switch 1"
     REMOTE = "WoRemote"
-
-
-class LockStatus(Enum):
-    LOCKED = 0
-    UNLOCKED = 1
-    LOCKING = 2
-    UNLOCKING = 3
-    LOCKING_STOP = 4  # LOCKING_BLOCKED
-    UNLOCKING_STOP = 5  # UNLOCKING_BLOCKED
-    NOT_FULLY_LOCKED = 6  # LATCH_LOCKED - Only EU lock type
